feat: extract core package with shared schemas, types, defaults, and constants
