+++ packages/core/src/index.ts
@@ -1,8 +1,25 @@
 /**
- * @module config/schemas
- * Re-exports all schemas from core.
+ * Core schemas, types, defaults, and identity constants for the jeeves-watcher ecosystem.
+ *
+ * @packageDocumentation
  */
 
+export {
+  COMPONENT_NAME,
+  CORE_PACKAGE,
+  DEFAULT_PORT,
+  PLUGIN_PACKAGE,
+  SERVICE_PACKAGE,
+} from './constants.js';
+export {
+  API_DEFAULTS,
+  CONFIG_WATCH_DEFAULTS,
+  EMBEDDING_DEFAULTS,
+  INIT_CONFIG_TEMPLATE,
+  LOGGING_DEFAULTS,
+  ROOT_DEFAULTS,
+  WATCH_DEFAULTS,
+} from './defaults.js';
 export {
   type ApiConfig,
   apiConfigSchema,
@@ -24,4 +41,4 @@ export {
   vectorStoreConfigSchema,
   type WatchConfig,
   watchConfigSchema,
-} from '@karmaniverous/jeeves-watcher-core';
+} from './schemas.js';
