+++ packages/core/src/schemas.ts
@@ -1,0 +1,7 @@
+/**
+ * @module schemas
+ * Re-exports all schemas from the modular schema structure.
+ * This file maintains backward compatibility while schemas are organized into logical groups.
+ */
+
+export * from './schemas/index.js';
