+++ packages/openclaw/package.json
@@ -49,20 +49,12 @@
       "./dist/index.js"
     ]
   },
-  "auto-changelog": {
-    "output": "CHANGELOG.md",
-    "tagPrefix": "openclaw/",
-    "unreleased": true,
-    "commitLimit": false,
-    "hideCredit": true
-  },
   "devDependencies": {
     "@dotenvx/dotenvx": "^1.59.1",
     "@rollup/plugin-commonjs": "^29.0.2",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@rollup/plugin-typescript": "^12.3.0",
-    "auto-changelog": "^2.5.0",
     "cross-env": "^10.1.0",
     "knip": "^6.3.0",
     "release-it": "^19.2.4",
@@ -75,7 +67,7 @@
     "build:skills": "tsx scripts/build-skills.ts",
     "build": "npm run build:plugin && npm run build:skills && npm run build:content",
     "build:content": "node scripts/copy-content.mjs",
-    "changelog": "auto-changelog",
+    "changelog": "git-cliff -o CHANGELOG.md",
     "lint": "eslint .",
     "lint:fix": "eslint --fix .",
     "release": "dotenvx run -f .env.local -- release-it",
@@ -85,7 +77,7 @@
   },
   "release-it": {
     "git": {
-      "changelog": "npx auto-changelog --unreleased-only --stdout --template https://raw.githubusercontent.com/release-it/release-it/main/templates/changelog-compact.hbs",
+      "changelog": "npx git-cliff --unreleased --strip header",
       "commitMessage": "chore: release @karmaniverous/jeeves-watcher-openclaw v${version}",
       "tagName": "openclaw/${version}",
       "requireBranch": "main"
@@ -94,20 +86,23 @@
       "release": true
     },
     "hooks": {
-      "after:bump": [
-        "npx auto-changelog -p",
-        "git add CHANGELOG.md"
-      ],
       "after:init": [
         "npm run lint",
         "npm run typecheck",
         "npm run test",
         "npm run build"
       ],
+      "before:npm:release": [
+        "npx git-cliff -o CHANGELOG.md",
+        "git add -A"
+      ],
       "after:release": [
         "git switch -c release/openclaw/${version}",
         "git push -u origin release/openclaw/${version}",
         "git switch ${branchName}"
+      ],
+      "after:bump": [
+        "node -e \"const f='openclaw.plugin.json';const j=JSON.parse(require('fs').readFileSync(f,'utf8'));j.version='${version}';require('fs').writeFileSync(f,JSON.stringify(j,null,2)+'')\""
       ]
     },
     "npm": {
@@ -115,6 +110,7 @@
     }
   },
   "dependencies": {
-    "@karmaniverous/jeeves": "^0.5.9"
+    "@karmaniverous/jeeves": "^0.5.9",
+    "@karmaniverous/jeeves-watcher-core": "*"
   }
 }
