+++ packages/service/src/config/index.ts
@@ -1,4 +1,21 @@
+/**
+ * @module config
+ * Configuration loading and re-exports of schemas and types from core.
+ */
+
 export { loadConfig } from './loadConfig';
+export type {
+  ApiConfig,
+  ConfigWatchConfig,
+  EmbeddingConfig,
+  InferenceRule,
+  JeevesWatcherConfig,
+  JeevesWatcherConfigInput,
+  LoggingConfig,
+  SchemaEntry,
+  VectorStoreConfig,
+  WatchConfig,
+} from '@karmaniverous/jeeves-watcher-core';
 export {
   apiConfigSchema,
   configWatchConfigSchema,
@@ -8,15 +25,4 @@ export {
   loggingConfigSchema,
   vectorStoreConfigSchema,
   watchConfigSchema,
-} from './schemas';
-export type {
-  ApiConfig,
-  ConfigWatchConfig,
-  EmbeddingConfig,
-  InferenceRule,
-  JeevesWatcherConfig,
-  JeevesWatcherConfigInput,
-  LoggingConfig,
-  VectorStoreConfig,
-  WatchConfig,
-} from './types';
+} from '@karmaniverous/jeeves-watcher-core';
